+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/AnimeNavDestinations.kt
@@ -39,4 +39,5 @@ enum class AnimeNavDestinations(val id: String) {
 
     USER_FOLLOWING("anime_user_following"),
     USER_FOLLOWERS("anime_user_followers"),
+    USER_FAVORITE_MEDIA("anime_user_favorite_media"),
 }