+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/forum/thread/comment/ForumThreadCommentTreeScreen.kt
@@ -129,7 +129,7 @@ object ForumThreadCommentTreeScreen {
                                 UpIconButton(upIconOption)
                             }
                         },
-                        colors = TopAppBarDefaults.centerAlignedTopAppBarColors(
+                        colors = TopAppBarDefaults.topAppBarColors(
                             containerColor = MaterialTheme.colorScheme.surfaceColorAtElevation(
                                 lerp(0.dp, 16.dp, scrollBehavior.state.overlappedFraction)
                             )