+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/history/MediaHistoryScreen.kt
@@ -93,7 +93,7 @@ object MediaHistoryScreen {
                                     )
                                 }
                             },
-                            colors = TopAppBarDefaults.centerAlignedTopAppBarColors(
+                            colors = TopAppBarDefaults.topAppBarColors(
                                 containerColor = MaterialTheme.colorScheme.surfaceColorAtElevation(
                                     lerp(0.dp, 16.dp, scrollBehavior.state.overlappedFraction)
                                 )