+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/home/AnimeHomeScreen.kt
@@ -272,7 +272,7 @@ object AnimeHomeScreen {
                                 )
                             }
                         },
-                        colors = TopAppBarDefaults.centerAlignedTopAppBarColors(
+                        colors = TopAppBarDefaults.topAppBarColors(
                             containerColor = MaterialTheme.colorScheme.surfaceColorAtElevation(
                                 lerp(0.dp, 16.dp, scrollBehavior.state.overlappedFraction)
                             )