+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/search/MediaSearchScreen.kt
@@ -215,7 +215,7 @@ object MediaSearchScreen {
                             }
                         }
                     },
-                    colors = TopAppBarDefaults.centerAlignedTopAppBarColors(
+                    colors = TopAppBarDefaults.topAppBarColors(
                         containerColor = MaterialTheme.colorScheme.surfaceColorAtElevation(
                             lerp(0.dp, 16.dp, scrollBehavior.state.overlappedFraction)
                         )