+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/user/follow/UserListScreen.kt
@@ -68,7 +68,7 @@ object UserListScreen {
                                     UpIconButton(option = upIconOption)
                                 }
                             },
-                            colors = TopAppBarDefaults.centerAlignedTopAppBarColors(
+                            colors = TopAppBarDefaults.topAppBarColors(
                                 containerColor = MaterialTheme.colorScheme.surfaceColorAtElevation(
                                     lerp(0.dp, 16.dp, scrollBehavior.state.overlappedFraction)
                                 )